+++ src/Api/Event/PuliEvents.php
@@ -20,16 +20,6 @@ namespace Puli\RepositoryManager\Api\Event;
 interface PuliEvents
 {
     /**
-     * Dispatched when a package file is loaded.
-     */
-    const LOAD_PACKAGE_FILE = 'puli.load-package-file';
-
-    /**
-     * Dispatched when a package file is saved.
-     */
-    const SAVE_PACKAGE_FILE = 'puli.save-package-file';
-
-    /**
      * Dispatched when the factory class is generated.
      */
     const GENERATE_FACTORY = 'puli.generate-factory';